+++ app/src/main/java/com/example/gogreen/data/StationData.kt
@@ -1,16 +1,16 @@
 package com.example.gogreen.data
 
 data class StationData(
-    val available_energy: Int,
-    val energy_rate: Int,
+    val available_energy: Double,
+    val energy_rate: Double,
     val energy_source: String,
     val energy_type: String,
     val id: String,
     val station_address: String,
     val station_name: String,
-    val total_cost: Int,
+    val total_cost: Double,
     val energy_unit: String,
-    val transferRate: Int,
-    val current: Int,
-    val voltage: Int
+    val transferRate: Double,
+    val current: Double,
+    val voltage: Double
 )
